--- a/src/Array.elm
+++ b/src/Array.elm
@@ -483,6 +483,6 @@ translateIndex idx arr =
         if posIndex < 0 then
             0
         else if posIndex > arr.length then
-            arr.length - 1
+            arr.length
         else
             posIndex