The current implementation of Arrays in elm is buggy, badly documented and hard to understand.

This PR introduces a re-implementation of Arrays which has none of the bugs of the current implementation, and is written using mostly Elm code.

## Implementation details

This implementation adds a new module that provides a thin, but immutable, wrapper over Javascript arrays called `JsArray`. This module has not been exposed, so it's only available within `core`. The module could be inlined into the `Array` module, but I felt it would be cleaner to have it as its own module as it deals with its own type.

The actual Array implementation is written in Elm and makes use of the `JsArray` module. It's not a Relaxed Radix Tree, which is why `append` operations are much slower.  On the other hand, this implementation "Just Works" with Elm's equality operator, which means that special handling of Arrays is unnecessary when performing equality checks. Checking two arrays for equality can also be an order of magnitude faster with this implementation.

`initialize` and `slice` operations are slower due to a naive implementation which could be improved in the future. On Chrome, `get` is about 50% of the performance, while in Safari `get` is on par with the current implementation. On both browsers, `set` and `push` are faster with this implementation.

Performance could improve with optimizations in the Elm language (like inlining Bitwise operations and wrapper-less function calls). Also, if a `pop` function is introduced, it could be used to improve performance of `slice` operations. Such a function would have to be careful to do the exact opposite of `push` to make sure `==` still works.

This implementation doesn't have any of the bugs the plagues the current Array implementation in `core`.

The implementation is based on the same idea as Persistent Vectors in Clojure. `push` pushes elements into a `tail`. When the tail is of size 32, a new empty tail is created, while the "full" tail is pushed into a HAMT-based tree. Pushing elements onto the array, as well as reading and replacing at the tail, is therefore a constant time operation.

To find a value in the tree, we divide the index into groups of five bits. This allows a single index to represent 7 groups of numbers between 0 and 31. We can use this groups to search for a value in the tree, where each group corresponds to an index at each level of the tree.

Which group represents the root/top of the tree, depends on the maximum depth of the tree. If an array has less than 32 elements, there really is no tree, we only need to look at the tail. If the tail overflows however, this is pushed into the tree, and we will have a maximum depth of two. The top level of the tree will in this case be represented by the second group of bits, while the first five bits will represent the position in the lowest level of the tree. Should the tree have a maximum depth of 3 (more than 1024 elements), the third group of bits will represent the position at the root.

This way of ordering the elements, ensures that the value corresponding to index 0 is at the bottom left of the tree, while the value corresponding to the highest index not in the tail, will be at the bottom right. In other words, the values are sorted by their corresponding index.

## Benchmarks

The benchmarks are available in the `native` branch on http://www.github.com/Skinney/collections-ng.git.

The results, on a macbook air 2012, running on the latest version of Chrome, are as follows:

For the current implementation:
```
Starting Large Array suite.
bench.js:8501 Build x 88.14 ops/sec ±2.12% (48 runs sampled)
bench.js:8501 Repeat x 4,259 ops/sec ±0.86% (60 runs sampled)
bench.js:8501 Set x 833,365 ops/sec ±1.94% (61 runs sampled)
bench.js:8501 Push x 788,039 ops/sec ±1.68% (60 runs sampled)
bench.js:8501 Get x 7,959,420 ops/sec ±1.78% (59 runs sampled)
bench.js:8501 Append x 448,015 ops/sec ±1.84% (59 runs sampled)
bench.js:8501 Slice from end x 706,822 ops/sec ±1.93% (60 runs sampled)
bench.js:8501 Slice from both x 383,147 ops/sec ±1.68% (61 runs sampled)
bench.js:8501 Fold x 7,460 ops/sec ±0.61% (40 runs sampled)
bench.js:8501 Indexed Map x 3,382 ops/sec ±1.69% (60 runs sampled)
bench.js:8501 Indexed List x 517 ops/sec ±3.56% (56 runs sampled)
bench.js:8501 Equality x 5,563 ops/sec ±0.79% (32 runs sampled)
bench.js:8501 Equality fail x 5,535 ops/sec ±0.75% (32 runs sampled)
bench.js:8501 Equality worst case x 3,296 ops/sec ±0.80% (21 runs sampled)
bench.js:8503 Done with Large Array suite.
```

For the implementation in this PR:
```
Starting Large Array suite.
bench.js:8501 Build x 101 ops/sec ±3.51% (51 runs sampled)
bench.js:8501 Repeat x 107 ops/sec ±0.95% (53 runs sampled)
bench.js:8501 Set x 1,423,405 ops/sec ±0.67% (61 runs sampled)
bench.js:8501 Push x 1,306,977 ops/sec ±0.75% (60 runs sampled)
bench.js:8501 Get x 3,101,145 ops/sec ±0.74% (61 runs sampled)
bench.js:8501 Append x 107 ops/sec ±2.46% (54 runs sampled)
bench.js:8501 Slice from end x 102 ops/sec ±4.07% (52 runs sampled)
bench.js:8501 Slice from both x 104 ops/sec ±1.42% (52 runs sampled)
bench.js:8501 Fold x 3,425 ops/sec ±0.61% (22 runs sampled)
bench.js:8501 Indexed Map x 101 ops/sec ±2.24% (52 runs sampled)
bench.js:8501 Indexed List x 808 ops/sec ±2.93% (54 runs sampled)
bench.js:8501 Equality x 1,113,068 ops/sec ±1.06% (61 runs sampled)
bench.js:8501 Equality fail x 1,118,502 ops/sec ±0.65% (61 runs sampled)
bench.js:8501 Equality worst case x 292 ops/sec ±0.72% (58 runs sampled)
bench.js:8503 Done with Large Array suite.
```

As mentioned earlier, results will vary between different browsers.