@shmookey 2016-07-24
To save others the effort of calculating the side-by-side comparison:
```Build               15% faster
Repeat              3980% slower
Set                 71% faster
Push                66% faster
Get                 39% faster
Append              420000% slower
Slice from end      690000% slower
Slice from both     370000% slower
Fold                220% slower
Indexed Map         3300% slower
Indexed List        56% faster
```