@Skinney 2016-07-24
No worries! I'm glad you added the numbers. That being said, I would appreciate if you also added numbers for equality, as some of the tradeoff has been on equality vs append performance. ;)

Repeat, initialize, slice, map, filter, append etc. are all essentially using a naive `foldl push` based solution while the current implementation uses specialized native code. That explains some of the difference in performance, but also indicates that there is potential for optimization.

The focus has been on correctness and getting the basic operations (get, set and push) competitive in performance.