@shmookey 2016-07-24
Sure thing. My thoughts were initially that it doesn't matter if your equality is faster or slower because it's correct and the old one is not - but I suppose it's nice to be able to show that it's also hundreds of times faster! :smile: