@Skinney 2016-07-24
Equality of arrays have been fixed in newer versions of `elm-lang/core` by special handling of arrays in the `eq` function (converts elm arrays to js arrays, then compare element for element). So it currently works, it's just slower than it has to be.