@shmookey 2016-07-25
> Also, if a pop function is introduced, it could be used to improve performance of slice operations.

Could you elaborate on this at all? Not that I find myself slicing arrays very often, but the performance hit here is large, and that seems like as good a reason as any to add a new function. Presumably it needn't even be exposed?