@Skinney 2016-07-25
Right. So currently, the slice operation is essentially implemented as a `foldn push empty source`. In the case where you only want to slice from the end, it would be faster just to `pop` elements off the array (doesn't require you to rebuild the array).

The first concern is to have a good foundation in core. Optimizations can come later. I'll be happy to work further on performance once this is merged, if it gets merged.