@evancz 2016-07-25
> Optimizations can come later. I'll be happy to work further on performance **if this is merged**.

I don't think this is a good way to approach things. It seems pretty smart to find a way to make slice faster. It's almost 700,000x slower, so I think that's worth exploring. Maybe the result is "We tried A and B and here's why they both do not work." That's better than "Maybe it can be better, but let's merge it into core before thinking it through."

I'm open to giving `Skinney/elm-array-draft` the ability to use the native wrapper for internal use of 32 element arrays if you *need* that so people who already have code that uses `Array` can try it and see more easily. I don't think that's what you were saying though.