@Skinney 2016-07-25
What I'm worried about is spending time optimizing something, and then not have this implementation used. But I see your point about slice, and other operations, being so much slower that it warrants further inspection. I'll keep at it and see what I come up with.