@evancz 2016-07-25
> What I'm worried about is spending time optimizing something, and then not have this implementation used.

This is *always* a possible outcome. I tried to express that [here](https://groups.google.com/forum/#!topic/elm-dev/NTnd_SuIzJc). For example, [the List.take improvements](https://groups.google.com/forum/#!topic/elm-dev/He_emRjsly8) started with something that was not really viable. At that time, we had no idea if it would *become* viable. If it didn't, wouldn't it have been a bad outcome to merge it?

Point is, thinking that "contribution" *means* "merged into core" is a poisonous mindset. Languages cannot work that way. Like I say [here](http://tech.noredink.com/post/147676730053/maybe-the-cookies-api-should-not-exist), many of *my* projects don't make the cut. Learning is progress. Prototypes are progress. Performance numbers are progress. Discussion is progress. Doing work doesn't mean core should change.

(This is also why it is *very important* to talk about projects like this on elm-dev while they are in progress. That I am hearing about this for the first time as a PR frames our discussion as accept-or-reject which is not a reasonable first discussion of a large and complex change IMO.)