@evancz 2016-07-25
To come back to my initial comments [here](https://groups.google.com/forum/#!topic/elm-dev/NTnd_SuIzJc), I think it is important that we find people using `Array` in their projects and see what kind of impact this has on them. Like, if someone has a grid based game, does it still work fine?

That data is important for assessing the way things are now, and will make it clearer what else would be needed. (Having it be a separate package while we assess these things makes it possible to gather data without having an accept-or-reject discussion yet.)

So next steps should be finding `Array` users.