@Skinney 2016-07-25
I don't disagree with anything you're saying, but I think there are some misconceptions here that is worth clearing up.

1) This code is already available in a separate library, called CollectionsNg, which also contains code for Dict and Set with working equality. This has been posted to the elm-dev mailing list in the past. The only difference between this PR and CollectionsNg (currently), is the JsArray module (CollectionsNg is available in elm-package, so no native code).

2) I'm well aware that this code might not make the cut, and that's fine. For those who value correctness over performance, CollectionsNg is already available. The reason I would rather wait to do optimization on other functions in the API, is because you've already signalled that a 40% drop in `get` performance might be too much. If the people using Array today cannot tolerate a 40% drop, and we are not able to close that gap, spending time on optimizing other functions would seem like a waste of time, unless those optimizations will be available in CollectionsNg (all my ideas for optimizing other functions currently involve more use of native code).

Anyway. I agree that getting more information from people of the Array module is the logical next step, so the question becomes how we go about that. I've advertised CollectionsNg any way I know how, and I don't think the library sees much use.

Should I close this PR until we have more information from users of `Array`? Either way is fine by me. :)