@evancz 2016-07-25
Given that equality on `Set` and `Dict` is fixed in core, I think it makes sense to create a package focused primarily on an alternate `Array` implementation. In the spirit of obvious names, I would suggest calling it `Skinney/elm-array-exploration` and having it only expose `Array.Hamt`.

I can't find the RRBT version, but I have talked to the person working on it at Elm meetups in SF a lot. Maybe it could be included in this package as `Array.Rrbt`, making it easy for people to compare.

I'm also down to let a library like this have native access, with the explicit goal of figuring out how best to have a better version in core.

Once such a thing exists, we can be explicit about encouraging people to give it a try and see how it works. Another way to get "user data" is to create projects that naturally want to use arrays and see how they look. Maybe certain domains do not use append often? Maybe others do? This kind of thing can be distributed easily, but we'll have to solicit help if we want it.

Does that route make sense?