@Skinney 2016-07-25
Sounds like a plan to me. I'll let you know when I have a repo up and running.