@Skinney 2016-07-30
@evancz @mgold manually inlining the bitwise operations had a pretty nice effect on performance.

Before inline:
```
Starting Large Array suite.
bench.js:7347 Build x 128 ops/sec ±2.57% (53 runs sampled)
bench.js:7347 Repeat x 775 ops/sec ±2.36% (51 runs sampled)
bench.js:7347 Set x 1,256,978 ops/sec ±2.36% (59 runs sampled)
bench.js:7347 Push x 1,249,187 ops/sec ±2.13% (59 runs sampled)
bench.js:7347 Get x 3,078,710 ops/sec ±0.76% (56 runs sampled)
bench.js:7347 Append x 109 ops/sec ±2.45% (50 runs sampled)
bench.js:7347 Slice from end x 103 ops/sec ±2.34% (52 runs sampled)
bench.js:7347 Slice from both x 104 ops/sec ±1.65% (52 runs sampled)
bench.js:7347 Fold x 3,455 ops/sec ±0.85% (22 runs sampled)
bench.js:7347 Indexed Map x 98.51 ops/sec ±2.43% (50 runs sampled)
bench.js:7347 Indexed List x 1,507 ops/sec ±0.95% (60 runs sampled)
bench.js:7347 Equality x 1,102,689 ops/sec ±0.90% (59 runs sampled)
bench.js:7347 Equality fail x 1,110,780 ops/sec ±1.62% (60 runs sampled)
bench.js:7347 Equality worst case x 290 ops/sec ±0.95% (56 runs sampled)
bench.js:7349 Done with Large Array suite.
```

After inline:
```
bench.js:7345 Starting Large Array suite.
bench.js:7347 Build x 128 ops/sec ±2.67% (53 runs sampled)
bench.js:7347 Repeat x 1,029 ops/sec ±2.44% (46 runs sampled)
bench.js:7347 Set x 1,531,744 ops/sec ±2.24% (58 runs sampled)
bench.js:7347 Push x 1,351,414 ops/sec ±2.31% (57 runs sampled)
bench.js:7347 Get x 5,004,991 ops/sec ±1.32% (58 runs sampled)
bench.js:7347 Append x 120 ops/sec ±2.16% (54 runs sampled)
bench.js:7347 Slice from end x 115 ops/sec ±2.09% (52 runs sampled)
bench.js:7347 Slice from both x 115 ops/sec ±1.86% (53 runs sampled)
bench.js:7347 Fold x 3,491 ops/sec ±0.80% (22 runs sampled)
bench.js:7347 Indexed Map x 110 ops/sec ±3.04% (50 runs sampled)
bench.js:7347 Indexed List x 1,554 ops/sec ±0.64% (12 runs sampled)
bench.js:7347 Equality x 1,111,154 ops/sec ±0.90% (61 runs sampled)
bench.js:7347 Equality fail x 1,118,650 ops/sec ±0.96% (61 runs sampled)
bench.js:7347 Equality worst case x 285 ops/sec ±0.99% (59 runs sampled)
bench.js:7349 Done with Large Array suite.
```